Fix build break Solaris/HPUX. isspace() wants type int.
git-svn-id: svn+ssh://source-control.beyondtrust.com/sources/repo/bt-dev/trunk/Platform/src/linux@48980 9fc1c11e-11b2-4e14-94d8-6c5e05775969

+++ lwreg/shell/regshell.c
@@ -1404,7 +1404,7 @@ DWORD RegShellSplitCmdParam(
     BAIL_ON_REG_ERROR(dwError);
 
     pszTmp = pszOutCmd;
-    while (*pszTmp && !isspace(*pszTmp))
+    while (*pszTmp && !isspace((int) *pszTmp))
     {
         pszTmp++;
     }
@@ -1414,7 +1414,7 @@ DWORD RegShellSplitCmdParam(
     if (dwCmdLen+1 < dwLen)
     {
         pszTmp++;
-        while (*pszTmp && isspace(*pszTmp))
+        while (*pszTmp && isspace((int) *pszTmp))
         {
             pszTmp++;
         }
